# Optimization for power-of-two-sized buffers