--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module github.com/balena-os/circbuf
+
+go 1.16